refactor(schedule): extract components + hooks for schedule page content
