+++ apps/web/src/components/dashboard/team/schedule/schedule-footer.tsx
@@ -1,0 +1,15 @@
+'use client'
+
+import { Button } from '@fuku/ui/components'
+import { Check } from 'lucide-react'
+
+export const ScheduleFooter = () => {
+  return (
+    <div className='flex gap-2 justify-end'>
+      <Button disabled>
+        Save
+        <Check />
+      </Button>
+    </div>
+  )
+}
