+++ apps/web/src/hooks/schedule/useScheduleData.tsx
@@ -1,0 +1,244 @@
+import { useParams } from 'next/navigation'
+import {
+  GenerateScheduleOutput,
+  RuleConditionCreateInput,
+  RuleConditionOutput,
+  RuleConditionUpdateInput,
+  RuleCreateInput,
+  RuleOutput,
+  RuleUpdateInput,
+} from '@fuku/api/schemas'
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
+import { DateTime } from 'luxon'
+
+import { useScheduleStore } from '~/store/schedule.store'
+import { useTRPC } from '~/trpc/client'
+
+export const useScheduleData = (start: Date, end: Date) => {
+  const params = useParams()
+  const slug = params.slug as string
+  const trpc = useTRPC()
+  const queryClient = useQueryClient()
+
+  const { setSchedulerAssignments } = useScheduleStore()
+
+  // API queries
+
+  const { data: team } = useQuery({
+    ...trpc.team.bySlug.queryOptions({ slug: slug! }),
+    enabled: !!slug,
+  })
+
+  const { data: teamMembers } = useQuery({
+    ...trpc.teamMember.list.queryOptions({ teamId: team?.id ?? '' }),
+    enabled: !!team,
+  })
+
+  const { data: shiftTypes } = useQuery({
+    ...trpc.shiftType.list.queryOptions({ teamId: team?.id ?? '' }),
+    enabled: !!team,
+  })
+
+  const { data: payGrades } = useQuery({
+    ...trpc.payGrade.list.queryOptions({ teamId: team?.id ?? '' }),
+    enabled: !!team,
+  })
+
+  const { data: dbAssignments } = useQuery({
+    ...trpc.assignment.list.queryOptions({
+      teamId: team?.id ?? '',
+      start,
+      end,
+    }),
+    enabled: !!team,
+  })
+
+  const { data: rules } = useQuery({
+    ...trpc.rule.groupById.queryOptions({ teamId: team?.id ?? '' }),
+    enabled: !!team,
+  })
+
+  const { data: ruleConditions } = useQuery({
+    ...trpc.ruleCondition.groupByRules.queryOptions({
+      teamId: team?.id ?? '',
+    }),
+    enabled: !!team,
+  })
+
+  // API mutations
+
+  const { mutateAsync: createRule } = useMutation({
+    ...trpc.rule.create.mutationOptions(),
+    onSuccess: data => {
+      queryClient.setQueryData(
+        trpc.rule.groupById.queryKey({ teamId: team?.id ?? '' }),
+        (oldData: Record<string, RuleOutput> | undefined) => {
+          if (!oldData) return oldData
+          return { ...oldData, [data.id]: data }
+        },
+      )
+    },
+  })
+
+  const { mutateAsync: updateRule } = useMutation({
+    ...trpc.rule.update.mutationOptions(),
+    onSuccess: data => {
+      queryClient.setQueryData(
+        trpc.rule.groupById.queryKey({ teamId: team?.id ?? '' }),
+        (oldData: Record<string, RuleOutput> | undefined) => {
+          if (!oldData) return oldData
+          return { ...oldData, [data.id]: data }
+        },
+      )
+    },
+  })
+
+  const { mutateAsync: deleteRule } = useMutation({
+    ...trpc.rule.delete.mutationOptions(),
+    onSuccess: data => {
+      queryClient.setQueryData(
+        trpc.rule.groupById.queryKey({ teamId: team?.id ?? '' }),
+        (oldData: Record<string, RuleOutput> | undefined) => {
+          if (!oldData) return oldData
+          const newData = { ...oldData }
+          delete newData[data.id]
+          return newData
+        },
+      )
+    },
+  })
+
+  const { mutateAsync: createRuleCondition } = useMutation({
+    ...trpc.ruleCondition.create.mutationOptions(),
+    onSuccess: data => {
+      queryClient.setQueryData(
+        trpc.ruleCondition.groupByRules.queryKey({ teamId: team?.id ?? '' }),
+        oldData => {
+          if (!oldData) return oldData
+          const ruleConditions = oldData[data.ruleId] ?? []
+          return {
+            ...oldData,
+            [data.ruleId]: [...ruleConditions, data],
+          }
+        },
+      )
+    },
+  })
+
+  const { mutateAsync: updateRuleCondition } = useMutation({
+    ...trpc.ruleCondition.update.mutationOptions(),
+    onSuccess: data => {
+      queryClient.setQueryData(
+        trpc.ruleCondition.groupByRules.queryKey({ teamId: team?.id ?? '' }),
+        oldData => {
+          if (!oldData) return oldData
+          const ruleConditions = oldData[data.ruleId] ?? []
+          const updatedRuleConditions = ruleConditions.map(rc =>
+            rc.id === data.id ? data : rc,
+          )
+          return { ...oldData, [data.ruleId]: updatedRuleConditions }
+        },
+      )
+    },
+  })
+
+  const { mutateAsync: deleteRuleCondition } = useMutation({
+    ...trpc.ruleCondition.delete.mutationOptions(),
+    onSuccess: data => {
+      queryClient.setQueryData(
+        trpc.ruleCondition.groupByRules.queryKey({ teamId: team?.id ?? '' }),
+        oldData => {
+          if (!oldData) return oldData
+          const ruleConditions = oldData[data.ruleId] ?? []
+          const updatedRuleConditions = ruleConditions.filter(
+            rc => rc.id !== data.id,
+          )
+          return { ...oldData, [data.ruleId]: updatedRuleConditions }
+        },
+      )
+    },
+  })
+
+  // Unified handlers for rules and rule conditions for RulePanelPopoverButton
+
+  const handleMutateRule = async (
+    mode: 'create' | 'update' | 'delete',
+    rule: RuleCreateInput | RuleUpdateInput | string,
+  ): Promise<RuleOutput> => {
+    switch (mode) {
+      case 'create':
+        return await createRule(rule as RuleCreateInput)
+      case 'update':
+        return await updateRule(rule as RuleUpdateInput)
+      case 'delete':
+        return await deleteRule({ id: rule as string })
+    }
+  }
+
+  const handleMutateRuleCondition = async (
+    mode: 'create' | 'update' | 'delete',
+    ruleCondition: RuleConditionCreateInput | RuleConditionUpdateInput | string,
+  ): Promise<RuleConditionOutput> => {
+    switch (mode) {
+      case 'create':
+        return await createRuleCondition(
+          ruleCondition as RuleConditionCreateInput,
+        )
+      case 'update':
+        return await updateRuleCondition(
+          ruleCondition as RuleConditionUpdateInput,
+        )
+      case 'delete':
+        return await deleteRuleCondition({ id: ruleCondition as string })
+    }
+  }
+
+  // Schedule generation mutation
+
+  const { mutateAsync: generateSchedule, isPending: isGenerating } =
+    useMutation({
+      ...trpc.schedule.generate.mutationOptions(),
+      onSuccess: (data: GenerateScheduleOutput) => {
+        setSchedulerAssignments(data.assignments)
+      },
+    })
+
+  const handleGenerateSchedule = () => {
+    if (!team) return
+
+    const timeZone = 'UTC' // team.timeZone
+
+    const startUTC = DateTime.fromJSDate(start)
+      .setZone(timeZone, { keepLocalTime: true })
+      .startOf('day')
+      .toUTC()
+
+    const endUTC = DateTime.fromJSDate(end)
+      .setZone(timeZone, { keepLocalTime: true })
+      .endOf('day')
+      .toUTC()
+
+    generateSchedule({
+      teamId: team.id,
+      start: startUTC.toJSDate(),
+      end: endUTC.toJSDate(),
+      timeZone,
+    })
+  }
+
+  return {
+    team,
+    teamMembers,
+    shiftTypes,
+    payGrades,
+    rules,
+    ruleConditions,
+    generateSchedule,
+    handleGenerateSchedule,
+    isGenerating,
+    handleMutateRule,
+    handleMutateRuleCondition,
+  }
+}
+
+export type ScheduleData = ReturnType<typeof useScheduleData>
